refactor(anoncreds)!: combine anoncreds packages
