+++ demo/src/BaseAgent.ts
@@ -10,7 +10,6 @@ import {
   V1CredentialProtocol,
   V1ProofProtocol,
 } from '@credo-ts/anoncreds'
-import { AnonCredsRsModule } from '@credo-ts/anoncreds-rs'
 import { AskarModule } from '@credo-ts/askar'
 import {
   CheqdAnonCredsRegistry,
@@ -122,8 +121,6 @@ function getAskarAnonCredsIndyModules() {
     }),
     anoncreds: new AnonCredsModule({
       registries: [new IndyVdrAnonCredsRegistry(), new CheqdAnonCredsRegistry()],
-    }),
-    anoncredsRs: new AnonCredsRsModule({
       anoncreds,
     }),
     indyVdr: new IndyVdrModule({
