--- packages/anoncreds-rs/README.md
@@ -1,31 +1,0 @@
-<p align="center">
-  <br />
-  <img
-    alt="Credo Logo"
-    src="https://github.com/openwallet-foundation/credo-ts/blob/c7886cb8377ceb8ee4efe8d264211e561a75072d/images/credo-logo.png"
-    height="250px"
-  />
-</p>
-<h1 align="center"><b>Credo AnonCreds RS Module</b></h1>
-<p align="center">
-  <a
-    href="https://raw.githubusercontent.com/openwallet-foundation/credo-ts/main/LICENSE"
-    ><img
-      alt="License"
-      src="https://img.shields.io/badge/License-Apache%202.0-blue.svg"
-  /></a>
-  <a href="https://www.typescriptlang.org/"
-    ><img
-      alt="typescript"
-      src="https://img.shields.io/badge/%3C%2F%3E-TypeScript-%230074c1.svg"
-  /></a>
-    <a href="https://www.npmjs.com/package/@credo-ts/anoncreds-rs"
-    ><img
-      alt="@credo-ts/anoncreds-rs version"
-      src="https://img.shields.io/npm/v/@credo-ts/anoncreds-rs"
-  /></a>
-
-</p>
-<br />
-
-AnonCreds RS module for [Credo](https://github.com/openwallet-foundation/credo-ts.git).
