--- packages/anoncreds-rs/jest.config.ts
@@ -1,13 +1,0 @@
-import type { Config } from '@jest/types'
-
-import base from '../../jest.config.base'
-
-import packageJson from './package.json'
-
-const config: Config.InitialOptions = {
-  ...base,
-  displayName: packageJson.name,
-  setupFilesAfterEnv: ['./tests/setup.ts'],
-}
-
-export default config
