--- packages/anoncreds-rs/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "@credo-ts/anoncreds-rs",
-  "main": "build/index",
-  "types": "build/index",
-  "version": "0.4.2",
-  "private": true,
-  "files": [
-    "build"
-  ],
-  "license": "Apache-2.0",
-  "publishConfig": {
-    "access": "public"
-  },
-  "homepage": "https://github.com/openwallet-foundation/credo-ts/tree/main/packages/anoncreds-rs",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/openwallet-foundation/credo-ts",
-    "directory": "packages/anoncreds-rs"
-  },
-  "scripts": {
-    "build": "yarn run clean && yarn run compile",
-    "clean": "rimraf ./build",
-    "compile": "tsc -p tsconfig.build.json",
-    "prepublishOnly": "yarn run build",
-    "test": "jest"
-  },
-  "dependencies": {
-    "@credo-ts/anoncreds": "0.4.2",
-    "@credo-ts/core": "0.4.2",
-    "class-transformer": "^0.5.1",
-    "class-validator": "0.14.0",
-    "rxjs": "^7.2.0",
-    "tsyringe": "^4.8.0"
-  },
-  "devDependencies": {
-    "@hyperledger/anoncreds-nodejs": "^0.2.0-dev.5",
-    "@hyperledger/anoncreds-shared": "^0.2.0-dev.5",
-    "@types/ref-array-di": "^1.2.6",
-    "@types/ref-struct-di": "^1.1.10",
-    "reflect-metadata": "^0.1.13",
-    "rimraf": "^4.4.0",
-    "typescript": "~4.9.5"
-  },
-  "peerDependencies": {
-    "@hyperledger/anoncreds-shared": "^0.2.0-dev.5"
-  }
-}
