--- packages/anoncreds-rs/src/errors/AnonCredsRsError.ts
@@ -1,7 +1,0 @@
-import { AriesFrameworkError } from '@credo-ts/core'
-
-export class AnonCredsRsError extends AriesFrameworkError {
-  public constructor(message: string, { cause }: { cause?: Error } = {}) {
-    super(message, { cause })
-  }
-}
