--- packages/anoncreds-rs/src/index.ts
@@ -1,6 +1,0 @@
-// Services
-export * from './services'
-
-// Module
-export { AnonCredsRsModule } from './AnonCredsRsModule'
-export { AnonCredsRsModuleConfig } from './AnonCredsRsModuleConfig'
