--- packages/anoncreds-rs/tests/setup.ts
@@ -1,4 +1,0 @@
-import '@hyperledger/anoncreds-nodejs'
-import 'reflect-metadata'
-
-jest.setTimeout(120000)
