--- packages/anoncreds-rs/tsconfig.build.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "outDir": "./build"
-  },
-  "include": ["src/**/*"]
-}
