+++ packages/anoncreds/src/__tests__/AnonCredsModuleConfig.test.ts
@@ -1,5 +1,6 @@
 import type { AnonCredsRegistry } from '../services'
 
+import { anoncreds } from '../../tests/helpers'
 import { AnonCredsModuleConfig } from '../AnonCredsModuleConfig'
 
 describe('AnonCredsModuleConfig', () => {
@@ -8,6 +9,7 @@ describe('AnonCredsModuleConfig', () => {
 
     const config = new AnonCredsModuleConfig({
       registries: [registry],
+      anoncreds,
     })
 
     expect(config.registries).toEqual([registry])
