+++ packages/anoncreds/src/error/AnonCredsRsError.ts
@@ -1,0 +1,7 @@
+import { AnonCredsError } from './AnonCredsError'
+
+export class AnonCredsRsError extends AnonCredsError {
+  public constructor(message: string, { cause }: { cause?: Error } = {}) {
+    super(message, { cause })
+  }
+}
