+++ packages/anoncreds/src/error/index.ts
@@ -1,2 +1,3 @@
 export * from './AnonCredsError'
 export * from './AnonCredsStoreRecordError'
+export * from './AnonCredsRsError'
