+++ packages/anoncreds/tests/LocalDidResolver.ts
