+++ packages/anoncreds/tests/setup.ts
@@ -1,1 +1,3 @@
+import '@hyperledger/anoncreds-nodejs'
+
 jest.setTimeout(120000)
