+++ packages/askar/package.json
@@ -32,8 +32,8 @@
     "tsyringe": "^4.8.0"
   },
   "devDependencies": {
-    "@hyperledger/aries-askar-nodejs": "^0.2.0-dev.5",
-    "@hyperledger/aries-askar-shared": "^0.2.0-dev.5",
+    "@hyperledger/aries-askar-nodejs": "^0.2.0-dev.6",
+    "@hyperledger/aries-askar-shared": "^0.2.0-dev.6",
     "@types/bn.js": "^5.1.0",
     "@types/ref-array-di": "^1.2.6",
     "@types/ref-struct-di": "^1.1.10",
@@ -42,6 +42,6 @@
     "typescript": "~4.9.5"
   },
   "peerDependencies": {
-    "@hyperledger/aries-askar-shared": "^0.2.0-dev.5"
+    "@hyperledger/aries-askar-shared": "^0.2.0-dev.6"
   }
 }
