+++ packages/indy-sdk-to-askar-migration/package.json
@@ -30,12 +30,12 @@
     "@credo-ts/node": "0.4.2"
   },
   "devDependencies": {
-    "@hyperledger/aries-askar-nodejs": "^0.2.0-dev.5",
-    "@hyperledger/aries-askar-shared": "^0.2.0-dev.5",
+    "@hyperledger/aries-askar-nodejs": "^0.2.0-dev.6",
+    "@hyperledger/aries-askar-shared": "^0.2.0-dev.6",
     "rimraf": "^4.4.0",
     "typescript": "~4.9.5"
   },
   "peerDependencies": {
-    "@hyperledger/aries-askar-shared": "^0.2.0-dev.5"
+    "@hyperledger/aries-askar-shared": "^0.2.0-dev.6"
   }
 }
