+++ samples/extension-module/package.json
@@ -24,6 +24,6 @@
     "@credo-ts/askar": "*",
     "class-validator": "0.14.0",
     "rxjs": "^7.2.0",
-    "@hyperledger/aries-askar-nodejs": "^0.2.0-dev.5"
+    "@hyperledger/aries-askar-nodejs": "^0.2.0-dev.6"
   }
 }
